Refactor: Move auth and files modules to 'modules' folder
- Moved auth and files modules to a new folder called 'modules'
- Updated imports in app.module.ts to reflect the new location
- Deleted now
