+++ src/app.module.ts
@@ -6,10 +6,10 @@ import { User } from './users/model/users.model'
 import { RolesModule } from './roles/roles.module'
 import { Role } from './roles/model/roles.model'
 import { UserRoles } from './roles/model/user-roles.model'
-import { AuthModule } from './auth/auth.module'
+import { AuthModule } from './modules/auth/auth.module'
 import { PostsModule } from './posts/posts.module'
 import { Post } from './posts/model/posts.model'
-import { FilesModule } from './files/files.module'
+import { FilesModule } from './modules/files/files.module'
 import { ServeStaticModule } from '@nestjs/serve-static'
 import * as path from 'path'
 @Module({
