+++ src/modules/users/users.module.ts
@@ -5,8 +5,8 @@ import { User } from './model/users.model'
 import { SequelizeModule } from '@nestjs/sequelize'
 import { Role } from 'src/roles/model/roles.model'
 import { RolesModule } from 'src/roles/roles.module'
-import { AuthService } from 'src/auth/auth.service'
-import { AuthModule } from 'src/auth/auth.module'
+import { AuthService } from 'src/modules/auth/auth.service'
+import { AuthModule } from 'src/modules/auth/auth.module'
 import { UserRoles } from 'src/roles/model/user-roles.model'
 import { Post } from 'src/posts/model/posts.model'
